refactor: make uint32_t casts explicit in tx tests

+++ src/test/transaction_tests.cpp
@@ -106,7 +106,7 @@ BOOST_AUTO_TEST_CASE(tx_valid)
                     break;
                 }
 
-                COutPoint outpoint(uint256S(vinput[0].get_str()), vinput[1].get_int());
+                COutPoint outpoint(uint256S(vinput[0].get_str()), uint32_t(vinput[1].get_int()));
                 mapprevOutScriptPubKeys[outpoint] = ParseScript(vinput[2].get_str());
             }
 
@@ -175,7 +175,7 @@ BOOST_AUTO_TEST_CASE(tx_invalid)
                     fValid = false;
                     break;
                 }
-                COutPoint outpoint(uint256S(vinput[0].get_str()), vinput[1].get_int());
+                COutPoint outpoint(uint256S(vinput[0].get_str()), uint32_t(vinput[1].get_int()));
                 mapprevOutScriptPubKeys[outpoint] = ParseScript(vinput[2].get_str());
             }
             if (!fValid)
